Refactor context block execution
Eliminate duplicate code in file and describe blocks.
